fake_news_detector/requirements.txt
newspaper3k
nltk
ddgs
transformers
torch
